FrozenPermeability model added to fem_examples
