--- FrozenPermeability/defaults/misc.json
@@ -1,4 +1,0 @@
-{
-  "processes": 4,
-  "cleanup": true
-}
