--- FrozenPermeability/defaults/model.json
@@ -1,6 +1,0 @@
-{
-    "theta": 0.0,
-    "iU": 0.0,
-    "iV": 0.0,
-    "iW": 0.0
-}
