--- FrozenPermeability/defaults/simulation.json
@@ -1,10 +1,0 @@
-{
-  "default": {},
-  "simple": {
-      "iU": 0.0,
-      "iV": 0.0,
-      "iW": 0.0,
-      "theta": 0.0
-
-  }
-}
