--- FrozenPermeability/docs/mkdocs.yml
@@ -1,7 +1,0 @@
-site_name: FrozenPermeability
-use_directory_urls: true
-theme:
-  name: material
-
-nav:
-  - Home: index.md
