--- FrozenPermeability/resources/rotor.dxf
@@ -1,662 +1,0 @@
-  0
-SECTION
-  2
-HEADER
-  9
-$INSBASE
- 10
-0.0000
- 20
-0.0000
-  9
-$EXTMIN
- 10
--32.388959821554153
- 20
--32.387879592170471
-  9
-$EXTMAX
- 10
-32.388959821554153
- 20
-32.387879592170471
-  9
-$LIMMIN
- 10
--32.388959821554153
- 20
--32.387879592170471
-  9
-$LIMMAX
- 10
-32.388959821554153
- 20
-32.387879592170471
-  9
-$TEXTSTYLE
-  7
-STANDARD
-  9
-$CLAYER
-  8
-default
-  0
-ENDSEC
-  0
-SECTION
-  2
-TABLES
-  0
-TABLE
-  2
-LTYPE
- 70
-4948253
-  0
-LTYPE
-  2
-CONTINUOUS
- 70
-    64
-  3
-Solid line
- 72
-    65
- 73
-     0
- 40
-0.0
-  0
-ENDTAB
-  0
-TABLE
-  2
-LAYER
- 70
-     5
-  0
-LAYER
-  2
-default
- 70
-    64
- 62
-     7
-  6
-CONTINUOUS
-  0
-LAYER
-  2
-layer1
- 70
-    64
- 62
-     7
-  6
-CONTINUOUS
-  0
-ENDTAB
-  0
-TABLE
-  2
-STYLE
- 70
-     1
-  0
-STYLE
-  2
-STANDARD
- 70
-     0
- 40
-0.0
- 41
-1.0
- 50
-0.0
- 71
-     0
- 42
-0.2
-  3
-txt
-  4
-
-  0
-ENDTAB
-  0
-TABLE
-  2
-VIEW
- 70
-     0
-  0
-ENDTAB
-  0
-ENDSEC
-  0
-SECTION
-  2
-BLOCKS
-  0
-ENDSEC
-  0
-SECTION
-  2
-ENTITIES
-  0
-LINE
-  8
-layer1
- 10
-28.278476710979433
- 20
--10.741985612846149
- 30
-0.0
- 11
-25.24029326269239
- 21
--9.5878879850197052
- 31
-0.0
-  0
-LINE
-  8
-layer1
- 10
-28.278476710979437
- 20
-10.741985612846142
- 30
-0.0
- 11
-25.240293262692393
- 21
-9.5878879850196981
- 31
-0.0
-  0
-LINE
-  8
-layer1
- 10
--25.24029326269239
- 20
-9.5878879850197034
- 30
-0.0
- 11
--28.278476710979433
- 21
-10.741985612846149
- 31
-0.0
-  0
-LINE
-  8
-layer1
- 10
--25.240293262692393
- 20
--9.5878879850196963
- 30
-0.0
- 11
--28.278476710979437
- 21
--10.741985612846142
- 31
-0.0
-  0
-LINE
-  8
-layer1
- 10
--4.8364059276779923
- 20
-29.860872018457883
- 30
-0.0
- 11
--4.3167920676795308
- 21
-26.652679156970674
- 31
-0.0
-  0
-LINE
-  8
-layer1
- 10
--23.442070783301421
- 20
-19.118886405611743
- 30
-0.0
- 11
--20.923501195012843
- 21
-17.064791171950983
- 31
-0.0
-  0
-LINE
-  8
-layer1
- 10
-4.3167920676795326
- 20
--26.652679156970674
- 30
-0.0
- 11
-4.8364059276779923
- 21
--29.860872018457883
- 31
-0.0
-  0
-LINE
-  8
-layer1
- 10
-20.923501195012843
- 20
--17.064791171950983
- 30
-0.0
- 11
-23.442070783301421
- 21
--19.118886405611743
- 31
-0.0
-  0
-LINE
-  8
-layer1
- 10
--23.442070783301443
- 20
--19.118886405611722
- 30
-0.0
- 11
--20.923501195012861
- 21
--17.064791171950958
- 31
-0.0
-  0
-LINE
-  8
-layer1
- 10
--4.8364059276780242
- 20
--29.860872018457883
- 30
-0.0
- 11
--4.3167920676795593
- 21
--26.652679156970677
- 31
-0.0
-  0
-LINE
-  8
-layer1
- 10
-20.923501195012857
- 20
-17.064791171950962
- 30
-0.0
- 11
-23.442070783301443
- 21
-19.118886405611722
- 31
-0.0
-  0
-LINE
-  8
-layer1
- 10
-4.316792067679561
- 20
-26.652679156970677
- 30
-0.0
- 11
-4.8364059276780242
- 21
-29.860872018457883
- 31
-0.0
-  0
-ARC
-  8
-layer1
- 10
--3.5527136788005009e-015
- 20
-1.7763568394002505e-015
- 30
-0
- 40
-30.249999999999996
- 50
-200.80000000000001
- 51
-219.19999999999999
-  0
-ARC
-  8
-layer1
- 10
-1.0658141036401503e-014
- 20
-3.5527136788005009e-015
- 30
-0
- 40
-30.250000000000007
- 50
-219.19999999999993
- 51
-260.79999999999995
-  0
-ARC
-  8
-layer1
- 10
--1.5987211554602254e-014
- 20
-3.5527136788005009e-015
- 30
-0
- 40
-30.250000000000004
- 50
-260.79999999999995
- 51
-279.19999999999999
-  0
-ARC
-  8
-layer1
- 10
--1.0658141036401503e-014
- 20
-1.4210854715202004e-014
- 30
-0
- 40
-30.250000000000018
- 50
-279.19999999999999
- 51
-320.79999999999995
-  0
-ARC
-  8
-layer1
- 10
--3.5527136788005009e-015
- 20
-1.4210854715202004e-014
- 30
-0
- 40
-30.250000000000011
- 50
-320.79999999999995
- 51
-339.19999999999999
-  0
-ARC
-  8
-layer1
- 10
-0
- 20
-1.7763568394002505e-015
- 30
-0
- 40
-30.249999999999996
- 50
-339.19999999999999
- 51
-20.79999999999999
-  0
-ARC
-  8
-layer1
- 10
-3.5527136788005009e-015
- 20
--1.7763568394002505e-015
- 30
-0
- 40
-30.249999999999996
- 50
-20.800000000000001
- 51
-39.199999999999974
-  0
-ARC
-  8
-layer1
- 10
--1.0658141036401503e-014
- 20
--3.5527136788005009e-015
- 30
-0
- 40
-30.250000000000007
- 50
-39.19999999999996
- 51
-80.799999999999955
-  0
-ARC
-  8
-layer1
- 10
-1.5987211554602254e-014
- 20
--3.5527136788005009e-015
- 30
-0
- 40
-30.250000000000004
- 50
-80.799999999999997
- 51
-99.200000000000017
-  0
-ARC
-  8
-layer1
- 10
-1.0658141036401503e-014
- 20
--1.4210854715202004e-014
- 30
-0
- 40
-30.250000000000018
- 50
-99.200000000000003
- 51
-140.79999999999995
-  0
-ARC
-  8
-layer1
- 10
-3.5527136788005009e-015
- 20
--1.4210854715202004e-014
- 30
-0
- 40
-30.250000000000011
- 50
-140.79999999999995
- 51
-159.19999999999999
-  0
-ARC
-  8
-layer1
- 10
-0
- 20
--1.7763568394002505e-015
- 30
-0
- 40
-30.249999999999996
- 50
-159.19999999999999
- 51
-200.79999999999998
-  0
-ARC
-  8
-layer1
- 10
-0
- 20
-1.7763568394002505e-015
- 30
-0
- 40
-27
- 50
-339.19999999999999
- 51
-20.79999999999999
-  0
-ARC
-  8
-layer1
- 10
-0
- 20
--1.7763568394002505e-015
- 30
-0
- 40
-27
- 50
-159.19999999999999
- 51
-200.79999999999998
-  0
-ARC
-  8
-layer1
- 10
--5.3290705182007514e-015
- 20
-7.1054273576010019e-015
- 30
-0
- 40
-26.999999999999993
- 50
-99.199999999999989
- 51
-140.80000000000001
-  0
-ARC
-  8
-layer1
- 10
-3.5527136788005009e-015
- 20
-0
- 30
-0
- 40
-27
- 50
-279.19999999999999
- 51
-320.79999999999995
-  0
-ARC
-  8
-layer1
- 10
-1.0658141036401503e-014
- 20
--3.5527136788005009e-015
- 30
-0
- 40
-27.000000000000004
- 50
-219.19999999999993
- 51
-260.79999999999995
-  0
-ARC
-  8
-layer1
- 10
--3.5527136788005009e-015
- 20
-1.0658141036401503e-014
- 30
-0
- 40
-26.999999999999993
- 50
-39.199999999999946
- 51
-80.799999999999955
-  0
-ARC
-  8
-layer1
- 10
-0
- 20
-0
- 30
-0
- 40
-12.35
- 50
-180
- 51
-360
-  0
-ARC
-  8
-layer1
- 10
-0
- 20
-0
- 30
-0
- 40
-12.35
- 50
-360
- 51
-180
-  0
-ENDSEC
-  0
-EOF
