+++ examples/fem_simulations/FrozenPermeability/defaults/model.json
@@ -1,4 +1,6 @@
 {
-  "x0": 1.0,
-  "mw": 5
+    "theta": 0.0,
+    "iU": 0.0,
+    "iV": 0.0,
+    "iW": 0.0
 }
