+++ examples/fem_simulations/FrozenPermeability/defaults/simulation.json
@@ -1,7 +1,10 @@
 {
-  "default": {
-    "t0": 0.0,
-    "t1": 5.3,
-    "nstep": 101
+  "default": {},
+  "simple": {
+      "iU": 0.0,
+      "iV": 0.0,
+      "iW": 0.0,
+      "theta": 0.0
+
   }
 }
