+++ examples/fem_simulations/FrozenPermeability/model.py
@@ -1,85 +1,75 @@
-import operator as op
-from copy import copy
-from itertools import cycle
-from math import acos, atan2, cos, degrees, fmod, hypot, pi, radians, sin
-
-import numpy as np
-
-from digital_twin_distiller import ModelPiece
+from operator import itemgetter
 from digital_twin_distiller.boundaries import DirichletBoundaryCondition
+from digital_twin_distiller.modelpiece import ModelPiece
 from digital_twin_distiller.material import Material
-from digital_twin_distiller.metadata import FemmMetadata
+from digital_twin_distiller.metadata import Agros2DMetadata, FemmMetadata
 from digital_twin_distiller.model import BaseModel
 from digital_twin_distiller.modelpaths import ModelDir
-from digital_twin_distiller.objects import Node
 from digital_twin_distiller.platforms.femm import Femm
+from digital_twin_distiller.platforms.agros2d import Agros2D
 from digital_twin_distiller.snapshot import Snapshot
+
+from itertools import cycle
+from math import cos, fmod, hypot, radians, sin
+from copy import copy
 
 ModelDir.set_base(__file__)
 
 
-def cart2pol(x: float, y: float):
-    r = hypot(x, y)
-    phi = degrees(atan2(y, x))
-    phi = fmod(phi + 1e-3, 360.0) - 1e-3
-    return r, phi
-
-
-def pol2cart(rho: float, phi: float):
-    x = rho * cos(radians(phi))
-    y = rho * sin(radians(phi))
-    return x, y
-
-
-def project(ax: float, ay: float, bx: float, by: float):
-    """
-    project vector 'a' onto vector 'b'.
-    """
-    c = ax * bx + ay * by / (bx**2 + by**2)
-    return c * bx, c * by
-
-
 class FrozenPermeability(BaseModel):
+    """this FrozenPermeability created when calling 'new'"""
+
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self._init_directories()
 
-        self.I0 = kwargs.get("I0", 0.0)
-        self.alpha = kwargs.get("alpha", 0.0)
-        self.rotorangle = kwargs.get("rotorangle", 0.0)
-
-        self.Nturns = 152
-        self.coil_area = 1
-        self.J0 = self.I0 * self.Nturns / self.coil_area
-
-        self.JU = self.J0 * cos(radians(self.alpha))
-        self.JV = self.J0 * cos(radians(self.alpha + 120))
-        self.JW = self.J0 * cos(radians(self.alpha + 240))
+        self.iU = kwargs.get("iU", 0.0)
+        self.iV = kwargs.get("iV", 0.0)
+        self.iW = kwargs.get("iW", 0.0)
+        self.theta = kwargs.get("theta", 0.0)
 
     def setup_solver(self):
+        # FEMM Solver
         femm_metadata = FemmMetadata()
         femm_metadata.problem_type = "magnetic"
         femm_metadata.coordinate_type = "planar"
         femm_metadata.file_script_name = self.file_solver_script
         femm_metadata.file_metrics_name = self.file_solution
         femm_metadata.unit = "millimeters"
-        femm_metadata.smartmesh = False
-        femm_metadata.depth = 1000  # TODO: Correct this value
+        femm_metadata.smartmesh = True
+        femm_metadata.depth = 30
+        platform_femm = Femm(femm_metadata)
 
-        self.platform = Femm(femm_metadata)
+        # Agros2D solver
+        agros_metadata = Agros2DMetadata()
+        agros_metadata.file_script_name = self.file_solver_script
+        agros_metadata.file_metrics_name = self.file_solution
+        agros_metadata.problem_type = "magnetic"
+        agros_metadata.coordinate_type = "planar"
+        agros_metadata.analysis_type = "steadystate"
+        agros_metadata.unit = 1e-3
+        agros_metadata.nb_refinements = 0
+        agros_metadata.adaptivity = "hp-adaptivity"
+        agros_metadata.adaptivity_tol = 1
+        agros_metadata.polyorder = 3
+        agros_metadata.adaptivity_steps = 50
+        platform_agros = Agros2D(agros_metadata)
+
+
+        # change the backend solver here
+        self.platform = platform_agros
+
         self.snapshot = Snapshot(self.platform)
 
     def define_materials(self):
-        # creating default materials
-        airgap = Material("airgap")
+        # creating material objects
         air = Material("air")
         steel_m19 = Material("M-19")
         steel_1018 = Material("1018")
         magnet = Material("Magnet")
         coil = Material("coil")
 
-        # default material configuration
-        airgap.meshsize = 0.1
+        # modifying material properties
 
         steel_m19.thickness = 0.635
         steel_m19.lamination_type = "inplane"
@@ -210,81 +200,56 @@ class FrozenPermeability(BaseModel):
         coil.lamination_type = "magnetwire"
         coil.conductivity = 58e6
 
-        # Generating additional materials
-        # with the above base materials
-
-        # Coils
-        # PHASE U
-        phase_U_positive = copy(coil)
-        phase_U_positive.name = "U+"
-        phase_U_positive.Je = self.JU
-
-        phase_U_negative = copy(coil)
-        phase_U_negative.name = "U-"
-        phase_U_negative.Je = -self.JU
-
-        # PHASE V
-        phase_V_positive = copy(coil)
-        phase_V_positive.name = "V+"
-        phase_V_positive.Je = self.JV
-
-        phase_V_negative = copy(coil)
-        phase_V_negative.name = "V-"
-        phase_V_negative.Je = -self.JV
-
-        # PHASE W
-        phase_W_positive = copy(coil)
-        phase_W_positive.name = "W+"
-        phase_W_positive.Je = self.JW
-
-        phase_W_negative = copy(coil)
-        phase_W_negative.name = "W-"
-        phase_W_negative.Je = -self.JW
-
-        # magnets
-        flipper = cycle((0, 180))
-        for i in range(6):
-            mi = copy(magnet)
-            mi.name = f"{mi.name}_{i+1}"
-            mi.remanence_angle = fmod(180 + i * 360 / 6 + next(flipper) + self.rotorangle, 360.0)
-            self.snapshot.add_material(mi)
-            label_x, label_y = pol2cart(28.5, i * 360 / 6 + self.rotorangle)
-            self.snapshot.assign_material(label_x, label_y, mi.name)
-
-        # coil labels
-        winding_labels = cycle(("U+", "W-", "V+", "U-", "W+", "V-"))
-        for i in range(18):
-            label_x, label_y = pol2cart(40, i * 360 / 18)
-            self.assign_material(label_x, label_y, next(winding_labels))
-
-        self.assign_material(0, 30.25 + 0.75 / 2, "airgap")
-
         self.snapshot.add_material(air)
-        self.snapshot.add_material(airgap)
         self.snapshot.add_material(steel_m19)
         self.snapshot.add_material(steel_1018)
-        self.snapshot.add_material(phase_U_positive)
-        self.snapshot.add_material(phase_U_negative)
-        self.snapshot.add_material(phase_V_positive)
-        self.snapshot.add_material(phase_V_negative)
-        self.snapshot.add_material(phase_W_positive)
-        self.snapshot.add_material(phase_W_negative)
+
+        flipper = cycle([0, 180])
+        for i in range(6):
+            # copy existing base material
+            newMagnet = copy(magnet)
+
+            # modify its properties
+            newMagnet.name = f"{newMagnet.name}_{i+1}"
+            newMagnet.remanence_angle = fmod(180 + i * 360 / 6 + next(flipper) + self.theta, 360.0)
+
+            # add it to the snapshot
+            self.snapshot.add_material(newMagnet)
+
+        coilArea = 0.000114934  # m2
+        N = 152  # number of turns
+        currentDensities = [N * self.iU / coilArea, N * self.iV / coilArea, N * self.iW / coilArea]
+        phaseNames = ["U", "V", "W"]
+        for i in range(3):
+            newPositiveCoilMaterial = copy(coil)
+            newPositiveCoilMaterial.name = phaseNames[i] + "+"
+            newPositiveCoilMaterial.Je = currentDensities[i]
+            self.snapshot.add_material(newPositiveCoilMaterial)
+
+            newNegativeCoilMaterial = copy(coil)
+            newNegativeCoilMaterial.name = phaseNames[i] + "-"
+            newNegativeCoilMaterial.Je = -currentDensities[i]
+            self.snapshot.add_material(newNegativeCoilMaterial)
 
     def define_boundary_conditions(self):
         a0 = DirichletBoundaryCondition("a0", field_type="magnetic", magnetic_potential=0.0)
 
-        # Adding boundary conditions to the snapshot
         self.snapshot.add_boundary_condition(a0)
 
         self.assign_boundary_arc(0, 53, "a0")
         self.assign_boundary_arc(0, -53, "a0")
 
     def add_postprocessing(self):
-        r = 30.25 + 0.75 / 2
-        phi = np.linspace(0, 2 * np.pi, 1001)
-        for phi_i in phi:
-            self.snapshot.add_postprocessing("point_value", (r * np.cos(phi_i), r * np.sin(phi_i)), "Bx")
-            self.snapshot.add_postprocessing("point_value", (r * np.cos(phi_i), r * np.sin(phi_i)), "By")
+        # filter the labels that lie inside a 30.5 mm radius circle.
+        # rotorLabels = [(x0, y0, "material0"), (x1, y1, "material1"), ... , (xn, yn, "materialn")]
+        rotorLabels = filter(lambda pi: hypot(pi[0], pi[1]) < 30.5, self.label_queue)
+
+        # drop the last element, the label name
+        # labelCoordinates = [(x0, y0), (x1,y1), ... , (xn, yn)]
+        labelCoordinates = map(itemgetter(0,1), rotorLabels)
+
+        # add torque calculation
+        self.snapshot.add_postprocessing("integration", labelCoordinates, "Torque")
 
     def build_geometry(self):
         stator = ModelPiece("stator")
@@ -292,74 +257,41 @@ class FrozenPermeability(BaseModel):
 
         rotor = ModelPiece("rotor")
         rotor.load_piece_from_dxf(ModelDir.RESOURCES / "rotor.dxf")
-        rotor.rotate(alpha=self.rotorangle)
+        rotor.rotate(alpha=self.theta)
 
-        self.geom.merge_geometry(stator.geom)
-        self.geom.merge_geometry(rotor.geom)
+        self.snapshot.add_geometry(stator.geom)
+        self.snapshot.add_geometry(rotor.geom)
 
-        for i in range(len(self.geom.circle_arcs)):
-            self.geom.circle_arcs[i].max_seg_deg = 1
-
-        self.snapshot.add_geometry(self.geom)
+        # Labels
 
         self.assign_material(0, 0, "air")
+        self.assign_material(0, 30.5, "air")
         self.assign_material(0, 50, "M-19")
         self.assign_material(0, 20, "1018")
 
-    def __call__(self, cleanup=True, devmode=False, timeout=10000):
-        r_ = super().__call__(cleanup=cleanup, devmode=devmode, timeout=timeout)
+        # coil labels
 
-        Bx = list(map(op.itemgetter(2), r_["Bx"]))
-        By = list(map(op.itemgetter(2), r_["By"]))
-        x = list(map(op.itemgetter(0), r_["Bx"]))
-        y = list(map(op.itemgetter(1), r_["By"]))
+        numberOfSlots = 18
+        windingLabels = cycle(("U+", "W-", "V+", "U-", "W+", "V-"))
+        vectorLength = 40
+        for i in range(numberOfSlots):
+            angle = radians(i * 360 / numberOfSlots)
+            labelX = vectorLength * cos(angle)
+            labelY = vectorLength * sin(angle)
+            self.assign_material(labelX, labelY, next(windingLabels))
 
-        res = {"Br": [], "Bphi": [], "r": [], "phi": []}
-        for Bxi, Byi, xi, yi in zip(Bx, By, x, y):
-            r_i, phi_i = cart2pol(xi, yi)
 
-            et = Node(xi, yi).rotate(radians(90))
-            et = et / abs(et)
-            er = Node.from_polar(1, phi_i)
+        # magnet labels
+        numberOfMagnets = 6
+        vectorLength = 28.5
+        for i in range(numberOfMagnets):
+            angle = radians(i * 360 / numberOfMagnets)
+            labelX = vectorLength * cos(angle)
+            labelY = vectorLength * sin(angle)
+            self.assign_material(labelX, labelY, f"Magnet_{i+1}")
 
-            # Itt kell kiszámítani a B(Bx, By) vektor vetületét a tangenciális (et) meg a radiális (er) irányú
-            # egységvektorokra
-            B_ri = project(Bxi, Byi, er.x, er.y)
-            B_ri = hypot(*B_ri)
-
-            B_phi = project(Bxi, Byi, et.x, et.y)
-            B_phi = hypot(*B_phi)
-
-            res["r"].append(r_i)
-            res["phi"].append(phi_i)
-            res["Br"].append(B_ri)
-            res["Bphi"].append(B_phi)
-
-        res["x"] = x
-        res["y"] = y
-        res["Bx"] = Bx
-        res["By"] = By
-
-        return res
 
 
 if __name__ == "__main__":
-    import json
-
-    import matplotlib.pyplot as plt
-
-    m = FrozenPermeability(I0=0.0, alpha=0.0, rotorangle=0.0, exportname="dev")
-    res_ = m(cleanup=False, devmode=False)
-
-    with open(ModelDir.DATA / "example.json", "w", encoding="utf-8") as f:
-        json.dump(res_, f, ensure_ascii=True, indent=2)
-
-    with open(ModelDir.DATA / "example.json", encoding="utf-8") as f:
-        res_ = json.load(f)
-
-    plt.figure(figsize=(10, 10))
-    plt.plot(res_["phi"][1:], res_["Bx"][1:])
-    plt.plot(res_["phi"][1:], res_["By"][1:])
-
-    plt.savefig(ModelDir.MEDIA / "example.png", bbox_inches="tight", dpi=450)
-    plt.show()
+    m = FrozenPermeability(exportname="dev", theta=0)
+    print(m(devmode=False, cleanup=False))
